Implement Undo(All)/Redo(All) logic
The logic now allows to undo redo any saved changes in the session and also has the undoAll redoAll variant to quickly get to the beginning or end of saved changes. Each step is memorized  by saving the file. This works only if the file is still open after saving. Closing the file will not allow to undo/redo previous changes when opened again.